requirements.txt
flask
psycopg2-binary
sqlalchemy
python-dotenv
flask-login
werkzeug
qrcode
pillow
psutil
requests
